requirements.txt
# dependencies
beautifulsoup4
colorama
md2pdf
python-dotenv
pyyaml
uvicorn
pydantic
fastapi
python-multipart
markdown
langchain
langchain_community
langchain-openai

tiktoken
gpt-researcher
arxiv
PyMuPDF
requests
jinja2
aiofiles
mistune
python-docx
htmldocx
lxml_html_clean
websockets
unstructured
json_repair
json5
loguru
langgraph==0.2.76

# uncomment for testing
# pytest
# pytest-asyncio

# 确保使用较新版本的OpenAI，它支持base_url参数
openai>=1.0.0
langchain-openai>=0.0.1
